requirements.txt
streamlit
transformers
sentence-transformers
PyPDF2
cohere
python-dotenv
